market_place/requirements.txt
annotated-types==0.5.0
anyio==4.0.0
blinker==1.6.2
certifi==2023.7.22
click==8.1.7
Flask==2.3.3
grpcio==1.58.0
grpcio-tools==1.58.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
nodeenv==1.8.0
packaging==23.1
pluggy==1.3.0
prisma==0.10.0
protobuf==4.24.3
pydantic==2.3.0
pydantic_core==2.6.3
pytest==7.4.2
python-dotenv==1.0.0
sniffio==1.3.0
tomlkit==0.12.1
typing_extensions==4.8.0
Werkzeug==2.3.7
